# Conflicts:
#	app/Http/Controllers/TagController.php
#	app/Http/Controllers/TopicController.php
#	app/Http/Controllers/VoteController.php

+++ app/Http/Controllers/AttachmentController.php
@@ -11,6 +11,7 @@ use App\Models\VoteItem;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use App\Facades\AttachmentService;
+use App\Http\Requests\AttachmentRequest;
 
 use App\Http\Requests;
 
@@ -53,13 +54,16 @@ class AttachmentController extends ApiController
         }
     }
 
+
     /**
      * @param Topic $topic
-     * @param Request $request
+     * @param AttachmentRequest $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function storeTopicAttachment(Topic $topic, Request $request)
+    public function storeTopicAttachment(Topic $topic, AttachmentRequest $request)
     {
+        $this->authorize('createTopicAttachment', $topic);
+
         $attachment_data = AttachmentService::uploadAttachmentToCloud($request);
         $attachment = Attachment::create($attachment_data);
         $attachment = $topic->attachments()->save($attachment);
@@ -75,11 +79,16 @@ class AttachmentController extends ApiController
     public function destroyTopicAttachment(Topic $topic, Attachment $attachment)
     {
         if ($this->isAttachmentBelongsToModel($topic, $attachment)) {
-            AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
-            $attachment->delete();
-            return $this->setStatusCode(204)->respond();
-        } else {
-            throw (new ModelNotFoundException)->setModel(Attachment::class);
+
+            $this->authorize('deleteTopicAttachment', $topic);
+
+            if ($this->isAttachmentBelongsToModel($topic, $attachment)) {
+                AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
+                $attachment->delete();
+                return $this->setStatusCode(204)->respond();
+            } else {
+                throw (new ModelNotFoundException)->setModel(Attachment::class);
+            }
         }
     }
 
@@ -111,11 +120,13 @@ class AttachmentController extends ApiController
 
     /**
      * @param Vote $vote
-     * @param Request $request
+     * @param AttachmentRequest $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function storeVoteAttachment(Vote $vote, Request $request)
+    public function storeVoteAttachment(Vote $vote, AttachmentRequest $request)
     {
+        $this->authorize('createVoteAttachment', $vote);
+
         $attachment_data = AttachmentService::uploadAttachmentToCloud($request);
         $attachment = Attachment::create($attachment_data);
         $attachment = $vote->attachments()->save($attachment);
@@ -131,11 +142,16 @@ class AttachmentController extends ApiController
     public function destroyVoteAttachment(Vote $vote, Attachment $attachment)
     {
         if ($this->isAttachmentBelongsToModel($vote, $attachment)) {
-            AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
-            $attachment->delete();
-            return $this->setStatusCode(204)->respond();
-        } else {
-            throw (new ModelNotFoundException)->setModel(Attachment::class);
+
+            $this->authorize('deleteVoteAttachment', $vote);
+
+            if ($this->isAttachmentBelongsToModel($vote, $attachment)) {
+                AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
+                $attachment->delete();
+                return $this->setStatusCode(204)->respond();
+            } else {
+                throw (new ModelNotFoundException)->setModel(Attachment::class);
+            }
         }
     }
 
@@ -165,13 +181,16 @@ class AttachmentController extends ApiController
         }
     }
 
+
     /**
      * @param Comment $comment
-     * @param Request $request
+     * @param AttachmentRequest $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function storeCommentAttachment(Comment $comment, Request $request)
+    public function storeCommentAttachment(Comment $comment, AttachmentRequest $request)
     {
+        $this->authorize('createCommentAttachment', $comment);
+
         $attachment_data = AttachmentService::uploadAttachmentToCloud($request);
         $attachment = Attachment::create($attachment_data);
         $attachment = $comment->attachments()->save($attachment);
@@ -187,11 +206,16 @@ class AttachmentController extends ApiController
     public function destroyCommentAttachment(Comment $comment, Attachment $attachment)
     {
         if ($this->isAttachmentBelongsToModel($comment, $attachment)) {
-            AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
-            $attachment->delete();
-            return $this->setStatusCode(204)->respond();
-        } else {
-            throw (new ModelNotFoundException)->setModel(Attachment::class);
+
+            $this->authorize('deleteCommentAttachment', $comment);
+
+            if ($this->isAttachmentBelongsToModel($comment, $attachment)) {
+                AttachmentService::deleteAttachmentFromCloud($attachment->cloud_public_id);
+                $attachment->delete();
+                return $this->setStatusCode(204)->respond();
+            } else {
+                throw (new ModelNotFoundException)->setModel(Attachment::class);
+            }
         }
     }
 
@@ -221,12 +245,13 @@ class AttachmentController extends ApiController
         }
     }
 
+
     /**
      * @param Message $message
-     * @param Request $request
+     * @param AttachmentRequest $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function storeMessageAttachment(Message $message, Request $request)
+    public function storeMessageAttachment(Message $message, AttachmentRequest $request)
     {
         $attachment_data = AttachmentService::uploadAttachmentToCloud($request);
         $attachment = Attachment::create($attachment_data);
