+++ app/Http/Controllers/CommentController.php
@@ -9,7 +9,7 @@ use App\Http\Requests\CommentsRequest;
 use App\Models\Topic;
 use App\Http\Requests;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-
+use Gate;
 class CommentController extends ApiController
 {
     /**
@@ -91,6 +91,8 @@ class CommentController extends ApiController
      */
     public function updateTopicComment(Topic $topic, Comment $comment, CommentsRequest $request)
     {
+        $this->authorize('updateTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)) {
             $comment->update($request->all());
             return $this->setStatusCode(200)->respond($comment);
@@ -107,6 +109,8 @@ class CommentController extends ApiController
      */
     public function destroyTopicComment(Topic $topic, Comment $comment)
     {
+        $this->authorize('deleteTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)) {
             $comment->delete();
             return $this->setStatusCode(204)->respond();
@@ -180,6 +184,9 @@ class CommentController extends ApiController
         Comment $commentChild,
         CommentsRequest $request
     ) {
+
+        $this->authorize('updateTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -199,6 +206,8 @@ class CommentController extends ApiController
      */
     public function destroyTopicCommentChild(Topic $topic, Comment $comment, Comment $commentChild)
     {
+        $this->authorize('deleteTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -218,7 +227,7 @@ class CommentController extends ApiController
      */
     protected function isCommentBelongsToVote(Vote $vote, Comment $comment)
     {
-        $voteWhichHasThisComment = $comment->votes()->get()->first();
+        $voteWhichHasThisComment = $comment->commentable()->get()->first();
 
         return ($voteWhichHasThisComment && $voteWhichHasThisComment->id === $vote->id);
     }
@@ -267,6 +276,8 @@ class CommentController extends ApiController
      */
     public function updateVoteComment(Vote $vote, Comment $comment, CommentsRequest $request)
     {
+        $this->authorize('updateVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)) {
             $comment->update($request->all());
             return $this->setStatusCode(200)->respond($comment);
@@ -283,6 +294,8 @@ class CommentController extends ApiController
      */
     public function destroyVoteComment(Vote $vote, Comment $comment)
     {
+        $this->authorize('deleteVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)) {
             $comment->delete();
             return $this->setStatusCode(204)->respond();
@@ -355,6 +368,8 @@ class CommentController extends ApiController
         Comment $commentChild,
         CommentsRequest $request
     ) {
+        $this->authorize('updateVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -374,6 +389,8 @@ class CommentController extends ApiController
      */
     public function destroyVoteCommentChild(Vote $vote, Comment $comment, Comment $commentChild)
     {
+        $this->authorize('deleteVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -393,7 +410,7 @@ class CommentController extends ApiController
      */
     protected function isCommentBelongsToVoteItem(VoteItem $voteItem, Comment $comment)
     {
-        $voteItemWhichHasThisComment = $comment->voteItems()->get()->first();
+        $voteItemWhichHasThisComment = $comment->commentable()->get()->first();
 
         return ($voteItemWhichHasThisComment && $voteItemWhichHasThisComment->id === $voteItem->id);
     }
@@ -442,6 +459,8 @@ class CommentController extends ApiController
      */
     public function updateVoteItemComment(VoteItem $voteItem, Comment $comment, CommentsRequest $request)
     {
+        $this->authorize('updateVoteItemsComment', [$comment, $voteItem]);
+
         if ($this->isCommentBelongsToVoteItem($voteItem, $comment)) {
             $comment->update($request->all());
             return $this->setStatusCode(200)->respond($comment);
@@ -458,6 +477,8 @@ class CommentController extends ApiController
      */
     public function destroyVoteItemComment(VoteItem $voteItem, Comment $comment)
     {
+        $this->authorize('deleteVoteItemsComment', [$comment, $voteItem]);
+
         if ($this->isCommentBelongsToVoteItem($voteItem, $comment)) {
             $comment->delete();
             return $this->setStatusCode(204)->respond();
